# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: AnthoLume API v1
  version: 1.0.0
  description: REST API for AnthoLume document management system
servers:
  - url: /api/v1
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        username:
          type: string
        is_admin:
          type: boolean
      required:
        - username
        - is_admin
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
      required:
        - code
        - message
...
    OperationType:
      type: string
      enum: [CREATE, UPDATE, DELETE]
    User:
      type: object
      properties:
        id:
          type: string
        admin:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
        - id
        - admin
        - created_at
    UsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
...
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: cookie
      name: token
paths:
...
  /auth/login:
    post:
      summary: User login
      operationId: login
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: Successful login
          headers:
            Set-Cookie:
              description: HttpOnly session cookie for authenticated requests.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/users:
    get:
      summary: Get all users
      operationId: getUsers
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create, update, or delete user
      operationId: updateUser
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                operation:
                  $ref: '#/components/schemas/OperationType'
                user:
                  type: string
                password:
                  type: string
                is_admin:
                  type: boolean
              required:
                - operation
                - user
      security:
        - BearerAuth: []
      responses:
        200:
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...